// apps/web/src/components/calendar/types.ts
export type TCalendarView = "day" | "week" | "month" | "year" | "agenda";
export type TEventColor =
	| "Blue"
	| "Green"
	| "Red"
	| "Yellow"
	| "Purple"
	| "Orange";

export type TGenres =
	| "Hip-Hop"
	| "Rock"
	| "Indie"
	| "Pop"
	| "Jazz"
	| "Classical"
	| "Electronic"
	| "Country"
	| "Reggae"
	| "Blues"
	| "Folk"
	| "Metal"
	| "R&B"
	| "Soul"
	| "Afrobeat"
	| "Punk"
	| "Disco"
	| "Funk"
	| "Gospel"
	| "Techno"
	| "House"
	| "Trance"
	| "Dubstep"
	| "Ambient"
	| "Alternative"
	| "Grunge"
	| "New Wave"
	| "Synthpop"
	| "Progressive Rock"
	| "Hard Rock"
	| "Soft Rock"
	| "Acoustic"
	| "Instrumental";